% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.R
\name{bins.fd}
\alias{bins.fd}
\title{Define number of bins for data discretization}
\usage{
bins.fd(values)
}
\arguments{
\item{values}{continuous data to discretize}
}
\value{
number of bins estimated through the Freedman–Diaconis rule
}
\description{
Calculates the number of bins for discretization according to the Freedman–Diaconis rule.
}
