% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.plotting.R
\name{create.data.type.colourlegend.heatmap}
\alias{create.data.type.colourlegend.heatmap}
\title{Create a data type pair heatmap}
\usage{
create.data.type.colourlegend.heatmap(
  data.type.pairs,
  data.type.colours,
  data.type.legend.labels
)
}
\arguments{
\item{data.type.pairs}{column names from the matrix of mutual information as output by calculate.mutual.information.pairwise.data.types.
This should be a character vector of the format <data.type.A>_<data.type.B> i.e. indicate the two variables for which mutual information was calculated delimted by '_'.
The vector should be ordered in a way consistent with other visualisations of the mutual information matrix e.g. by median mutual information of data type pairs.}

\item{data.type.colours}{a named character vector list mapping data types to desired colours. Names should match those in data.type.pairs}
}
\value{
a list with elements data.types.map and data.types.legend.
data.types.map is a heatmap providing a visual representation of the pariwise data type comparisons for which mutual information was calculated with calculate.mutual.information.pairwise.data.types.
data.types.legend is a legend visualising the data type to colour mapping in data.type.colours.
}
\description{
Create a heatmap providing a visual representation of the pariwise data type comparisons for which mutual information was calculated with calculate.mutual.information.pairwise.data.types, along with an accompanying legend.
}
