% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.plotting.R
\name{create.mutual.information.multipanelplot}
\alias{create.mutual.information.multipanelplot}
\title{Assemble mutual information mutlipanelplot}
\usage{
create.mutual.information.multipanelplot(
  MI.violinplot,
  data.types.heatmap,
  num.genes.heatmap,
  num.sig.genes.heatmap,
  perc.sig.genes.heatmap,
  plot.objects.heights = c(0.56, 0.14, 0.1, 0.1, 0.1),
  data.types.legend,
  num.genes.legend,
  num.sig.genes.legend,
  perc.sig.genes.legend,
  return.plot = FALSE,
  filename = NULL,
  plot.width = NULL,
  plot.height = NULL,
  y.spacing = -0.5,
  style = "Nature"
)
}
\arguments{
\item{MI.violinplot}{mutual information violinplot at returned by create.mutual.information.violinplot}

\item{num.genes.heatmap}{a heatmap displaying numbers of genes analysed in each pairwise data type comparison. As returned by create.mutual.information.summary.statistic.heatmap.}

\item{num.sig.genes.heatmap}{a heatmap displaying numbers of genes with significant mutual information in each pairwise data type comparison. As returned by create.mutual.information.summary.statistic.heatmap.}

\item{plot.objects.heights}{relative heights of each plot i.e. in decimal scale where the sum is equal to one.}

\item{data.types.legend}{a legend visualising the data type to colour mapping in data.type.colours. As returned by create.data.type.colourlegend.heatmap.}

\item{num.genes.legend}{a legend mapping numbers of genes displayed in num.genes.heatmap to colour scheme. As returned by create.mutual.information.summary.statistic.heatmap.}

\item{num.sig.genes.legend}{a legend mapping numbers of genes displayed in num.sig.genes.heatmap to colour scheme. As returned by create.mutual.information.summary.statistic.heatmap.}

\item{return.plot}{should the plot object be returned. Defaults to FALSE.}

\item{filename}{path to write plot to. If NULL no image is saved. Defaults to NULL.}

\item{plot.width}{if filename is not NULL, the width of the image to be save in inches. See ?create.multipanelplot.}

\item{plot.height}{if filename is not NULL, the height of the image to be save in inches. See ?create.multipanelplot.}

\item{y.spacing}{vertical spacing between each plot. See ?create.multipanelplot}

\item{style}{'BoutrosLab' or 'Nature'. Defaults to 'Nature'. See ?create.multipanelplot.}
}
\value{
a multipanelplot visualising the input plots and legends.
}
\description{
Combine all of the individual plots and legend into a single plot.
}
