% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.plotting.R
\name{create.mutual.information.summary.plot}
\alias{create.mutual.information.summary.plot}
\title{Create a summary plot from a mutual information analysis}
\usage{
create.mutual.information.summary.plot(
  MI,
  order.by.median = TRUE,
  log.scale = TRUE,
  data.type.colours,
  data.type.legend.labels = names(data.type.colours),
  summary.statistics,
  return.plot = FALSE,
  return.plot.elements = FALSE,
  filename = NULL,
  plot.width = ncol(MI),
  plot.height = 5,
  style = "Nature",
  violinplot.yaxis.cex = 0.55,
  violinplot.ylabel = "Mutual Information",
  violinplot.y.min.log = 0.01,
  violinplot.yaxis.lab = NULL,
  violinplot.yat = NULL,
  violinplot.ylimits = NULL
)
}
\arguments{
\item{MI}{matrix of mutual information as output by calculate.mutual.information.pairwise.data.types.
The column names should have the format <data.type.A>_<data.type.B> i.e. indicate the two variables for which mutual information was calculated delimted by '_'.
Rows correspond to the entities from which mutual information was calculated. i.e. genes.}

\item{order.by.median}{Should the data corresponding to each pair of data types analysed be displayed in order of median mutual information (highest to lowest). Defaults to TRUE.}

\item{log.scale}{plot mutual information on a log scale? Assumes mutual information is normalized (0)}

\item{data.type.colours}{a named character vector list mapping data types to desired colours. Names should match data types in colnames of `MI`.}

\item{summary.statistics}{matrix of MI summary statistics as returned by `calculate.mutual.information.pairwise.data.types`. Assumed to have row.names corresponding to colnames of `MI`.
Should have columns named 'num.genes' and 'num.sig.genes'.}

\item{return.plot}{should the plot object be returned. If return.plot.elements = TRUE, this argument is ignored. Defaults to FALSE.}

\item{return.plot.elements}{should individual plot elements be returned instead of assembling into a multipanelplot. If TRUE, a list with elements 'plots' and 'legends' is returned. Defaults to FALSE.}

\item{filename}{path to write plot to. If NULL no image is saved. If return.plot.elements = TRUE, this argument is ignored. Defaults to NULL.}

\item{plot.width}{if filename is not NULL, the width of the image to be saved in inches. See ?create.multipanelplot. Defaults to ncol(MI).}

\item{plot.height}{if filename is not NULL, the height of the image to be saved in inches. See ?create.multipanelplot. Defaults to 5.}

\item{style}{'BoutrosLab' or 'Nature'. Defaults to 'Nature'. See ?create.multipanelplot.}

\item{violinplot.yaxis.cex}{yaxis.cex for mutual information violinplot.}

\item{violinplot.ylabel}{ylabel for mutual information violinplot.}

\item{violinplot.y.min.log}{minimum value to display in violinplot if using a log10 scale. Defaults to 0.01}

\item{violinplot.yaxis.lab}{yaix.lab for mutual information violinplot.}

\item{violinplot.yat}{yat for for mutual information violinplot.}
}
\value{
Plots consist of:
A violinplot displaying the distribution of observed mutual information across genes between pairs of data types.
A heatmap providing a visual representation of the pairwise data type comparisons for which mutual information was calculated.
A heatmap displaying numbers of genes analysed in each pairwise data type comparison.
A heatmap displaying numbers of genes with significant mutual information in each pairwise data type comparison.
If return.plot.elements is TRUE, individual plot elements are returned.
If FALSE, plot elements are assembled into a multipanelplot.
If return.plot is TRUE, the multipanelplot is returned.
If return.plot.elements is FALSE and filename is not NULL, the multipanelplot is saved to a file.
}
\description{
Creates four plot elements and optionally assembles them into a multipanelplot or returns them individually.
}
\examples{
# See example code in ?calculate.mutual.information.pairwise.data.types

}
