% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.plotting.R
\name{create.mutual.information.summary.statistic.heatmap}
\alias{create.mutual.information.summary.statistic.heatmap}
\title{Create heatmap displaying summary statistic of interest}
\usage{
create.mutual.information.summary.statistic.heatmap(
  summary.statistic,
  statistic.name,
  colour.scheme
)
}
\arguments{
\item{summary.statistic}{vector containing summary statistic relating to each pairwise data type compraison.}

\item{colour.scheme}{vector of length two mapping values of summary statistic.
First element corresponds to colour used for 0 value, second element corresponds to colour used for max(summary.statistic) scheme}
}
\value{
a list with elements summary.statistic.heatmap summary.statistic.legend
summary.statistic.heatmap is a heatmap displaying summary.statistic.
summary.statistic.legend is a legend visualising the mapping of summary.statistic colour.scheme.
}
\description{
Create heatmap displaying summary statistic of interest for each pairwise data type comparison.
e.g. the total number of genes analysed in each pairwise data type comparison, or the number of those genes having significant mutual information can be displayed.
}
