% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.plotting.R
\name{create.mutual.information.violinplot}
\alias{create.mutual.information.violinplot}
\title{Create mutual information violinplot}
\usage{
create.mutual.information.violinplot(
  MI,
  log.scale,
  ylabel = "Mutual Information",
  yaxis.cex = 0.55,
  y.min.log = 0.01,
  style = "Nature",
  yaxis.lab = NULL,
  yat = NULL,
  ylimits = NULL
)
}
\arguments{
\item{MI}{matrix of mutual information as output by calculate.mutual.information.pairwise.data.types.
The column names should have the format <data.type.A>_<data.type.B> i.e. indicate the two variables for which mutual information was calculated delimted by '_'.
Rows correspond to the entities from which mutual information was calculated. i.e. genes.bounded between 0-1}

\item{log.scale}{plot mutual information on a log scale? Assumes mutual information is normalized (0)}

\item{ylabel}{label to use for y-axis . Defaults to 'Mutual Information').}

\item{yaxis.cex}{see ?create.violinplot}

\item{y.min.log}{minimum value to display if using a log10 scale. Defaults to 0.01}

\item{style}{BoutrosLab.plotting style, see ?create.violinplot for more. Defaults to 'Nature'.}

\item{yaxis.lab}{see ?create.violinplot}
}
\value{
a violinplot displaying the distribution of observed mutual information across genes between pairs of data types.
}
\description{
Designed for visualising the mutual information matrix returned by calculate.mutual.information.pairwise.data.types.
}
