% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.utils.R
\name{get.mutual.information.from.matrices}
\alias{get.mutual.information.from.matrices}
\title{Calculate mutual information between rows of two matrices}
\usage{
get.mutual.information.from.matrices(
  input.matrix.1,
  input.matrix.2,
  input.matrix.1.categorical = FALSE,
  input.matrix.2.categorical = FALSE,
  match.rows = TRUE,
  match.columns = TRUE,
  filter.rows = FALSE,
  filter.threshold.1,
  filter.threshold.2,
  normalize.input = FALSE,
  normalize.mutual.info = TRUE,
  n.permutations,
  calculate.pvalue = FALSE,
  n.cores = 1,
  n.bins,
  FD.n.bins = FALSE,
  sample.size.bins,
  equal.frequency.bins,
  use.shrinkage = FALSE,
  simple.sample.size.correction = FALSE
)
}
\arguments{
\item{input.matrix.1}{first input matrix from which to calculate mutual information}

\item{input.matrix.2}{second input matrix from which to calculate mutual information}

\item{input.matrix.1.categorical}{does input.matrix.2 represent a categorical variable. Used if FD.n.bins is TRUE. If so, the number of bins used is equal to the number of unique values for each row of input.matrix.2 separately.
input.matrix.2 must be numeric to allow downstream discretization, therefore categorical data must still be encoded in numeric form. Defaults to FALSE.}

\item{match.rows}{if TRUE, common row.names of input.matrices are identified and used for computation of mutual information. Defaults to TRUE.
If FALSE, input matrices should have the same number of rows, which will be compared in order i.e. the first row of input.matrix.1 compared with the first row of input.matrix.2. Defaults to TRUE.}

\item{match.columns}{if TRUE, columns of input matrices will be orderd by shared colnames (i.e. sample names). If FALSE, columns of input matrices should already be matched.}

\item{filter.rows}{whether to filter rows from each data type pair after intersecting by column names (samples). This may be desirable as data may become 'too sparse' after intersecting
between a given data type pair. Defaults to FALSE.}

\item{filter.threshold.1}{used if filter.rows is TRUE. What is the minimum proportion of samples that should have non-zero values for each row of input.matrix.1.
e.g. 0.5 would indicate >= 50% of samples must have non-zero observations.}

\item{filter.threshold.2}{used if filter.rows is TRUE. What is the minimum proportion of samples that should have non-zero values for each row of input.matrix.2.}

\item{normalize.input}{should data be normalized using the inverse normal transformation.
See `?RNOmni::RankNorm`. Normalization is performed after intersecting data types by sample and optional filtering of rows. This may be more desirable than normalizing data before intersection and sampling.
Defaults to FALSE}

\item{normalize.mutual.info}{should the mutual information be normalized (bounded between 0-1)}

\item{n.permutations}{if calculate.pvalue = TRUE, determines number of permutations to use}

\item{calculate.pvalue}{whether to calculate probability of mutual information using a permutation-based approach. Defaults to FALSE}

\item{n.cores}{how many cores to use for parallelization of mutual information calculation. Parallelization occurs across genes/rows of input matrices. Defaults to 1.}

\item{n.bins}{number of bins to use in discretizing data, applied to both input vectors. Used if FD.n.bins = FALSE}

\item{FD.n.bins}{whether to calculate number of bins for discretization based upon input data using the Freedman–Diaconis rule.
Performed for both input vectors separately. Overides n.bins and sample.size.bins. Performed for both input vectors separately. Defaults to FALSE}

\item{sample.size.bins}{whether to calculate number of bins for discretization as a function of sample size (N).
Bin number = N^(1/3). Defaults to FALSE}

\item{equal.frequency.bins}{whether to use an adaptive bin width which aims to equalize the frequencies of observations in each bin. Defaults to FALSE}

\item{use.shrinkage}{whether to perform shrinkage when calculating mutual information using the James-Stein estimator. Defaults to FALSE}

\item{simple.sample.size.correction}{apply a naive correction to reduce the bias coming from a limited sample size}
}
\value{
matrix containing mutual information for every matched row of input matrices.
If calculate.pvalue = TRUE, three columns are returned:
observed/true mutual information, expected mutual information = median of permutations, p.value = probability of observed mutual information
the number of rows will be the length of the intersections of the row.names between input matrices
}
\description{
Helper function - runs get.MI on two matrices with common rows (genes)
Matching of rows and colnames is optionally performed
Matching colnames (i.e. samples) and row.names (i.e. genes) are identified and used to order the two input matrices
}
