% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.R
\name{get.mutual.information}
\alias{get.mutual.information}
\title{Wrapper function - Calculate mutual information}
\usage{
get.mutual.information(
  data.1,
  data.2,
  data.1.categorical = FALSE,
  data.2.categorical = FALSE,
  n.bins,
  FD.n.bins = FALSE,
  equal.frequency.bins = FALSE,
  calculate.pvalue,
  n.permutations = 10^5,
  normalize.mutual.info = TRUE,
  use.shrinkage = FALSE,
  simple.sample.size.correction = FALSE
)
}
\arguments{
\item{data.1}{first input vector from which to calculate mutual information}

\item{data.2}{second input vector from which to calculate mutual information}

\item{data.1.categorical}{does data.1 represent a categorical variable. Used if FD.n.bins is TRUE. If so, the number of bins used is equal to the number of unique values of data.1.
data.1 must be numeric for discretization, therefore categorical data must still be encoded in numeric form. Defaults to FALSE.}

\item{data.2.categorical}{does data.1 represent a categorical variable. Used if FD.n.bins is TRUE. If so, the number of bins used is equal to the number of unique values of data.1.
data.1 must be numeric for discretization, therefore categorical data must still be encoded in numeric form. Defaults to FALSE.}

\item{n.bins}{number of bins to use in discretizing data, applied to both input vectors. Used if FD.n.bins = FALSE}

\item{FD.n.bins}{whether to calculate number of bins for discretization based upon input data using the Freedman–Diaconis rule.
Performed for both input vectors separately. Overides n.bins. Performed for both input vectors separately.
If the IQR for a given set of values is 0, the number of bins cannot be calculated and an NA value for mutual information will be returned.}

\item{equal.frequency.bins}{whether to use an adaptive bin width which aims to equalize the frequencies of observations in each bin. Defaults to FALSE}

\item{calculate.pvalue}{whether to calculate probability of mutual information using a permutation-based approach}

\item{n.permutations}{if calculate.pvalue = TRUE, determines number of permutations to use. Defaults to 10^5.}

\item{normalize.mutual.info}{should the mutual information be normalized (bounded between 0-1). Defaults to TRUE.}

\item{use.shrinkage}{whether to perform shrinkage when calculating mutual information using the James-Stein estimator. Defaults to FALSE}

\item{simple.sample.size.correction}{apply a naive correction to reduce the bias coming from a limited sample size}
}
\value{
If calculate.pvalue = FALSE, mutual information is returned (after optional normalization).
If calculate.pvalue = TRUE, a vector of length three is returned continaing:
observed/true mutual information, expected mutual information = median of purmutations, p.value = probability of observed mutual information
}
\description{
Calls mutual.information function. Handles cases with invariate data. Handles optional permutation-based significance estimation
}
