% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual.information.utils.R
\name{normalize.data}
\alias{normalize.data}
\title{Normalize data using the rank-based inverse normal transform.}
\usage{
normalize.data(input.matrix)
}
\arguments{
\item{input.matrix}{numeric matrix, rows of which will be normalized.}
}
\value{
normalized matrix
}
\description{
Rows of a matrix of normalized using the rank-based inverse normal transform.
See `?RNOmni::RankNorm`
}
