% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.data.R
\docType{data}
\name{simple.data}
\alias{simple.data}
\title{simple.data}
\format{
a list of 3 data.frames each with the same features in rows and sample sin columns:
\describe{
\item{data.a}{continuous features}
\item{data.b}{continuous features}
\item{data.c}{categorical features}
}
}
\usage{
simple.data
}
\description{
Simple dataset used for testing package.
}
\examples{
data('simple.data');
lapply(simple.data, dim);
lapply(simple.data, function(x) x[1:5,1:5]);

}
\keyword{datasets}
